+++ plugin/src/main/groovy/org/openbakery/tools/Xcode.groovy
@@ -17,6 +17,8 @@ class Xcode {
 	String xcodePath
 	Version version = null
 
+	HashMap<String, String> buildSettings = null
+
 
 	public Xcode(CommandRunner commandRunner) {
 		this(commandRunner, null)
@@ -104,6 +106,34 @@ class Xcode {
 		return getPath() + "/Contents/Developer/usr/bin/simctl"
 	}
 
+	String getToolchainDirectory() {
+		String buildSetting = getBuildSetting("TOOLCHAIN_DIR")
+		if (buildSetting != null) {
+			return buildSetting
+		}
+		return "/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain"
+	}
+
+	String loadBuildSettings() {
+		return commandRunner.runWithResult(getXcodebuild(), "-showBuildSettings")
+	}
+
+	String getBuildSetting(String key) {
+		if (buildSettings == null) {
+			buildSettings = new HashMap<>()
+			String[] buildSettingsData = loadBuildSettings().split("\n")
+			for (line in buildSettingsData) {
+				int index = line.indexOf("=")
+
+				String settingsKey = line.substring(0, index).trim()
+				String settingsValue = line.substring(index+1, line.length()).trim()
+
+				buildSettings.put(settingsKey, settingsValue)
+
+			}
+		}
+		return buildSettings.get(key)
+	}
 
 	@Override
 	public String toString() {