+++ plugin/src/test/groovy/org/openbakery/tools/XcodeSpecification.groovy
@@ -222,4 +222,24 @@ class XcodeSpecification extends Specification {
 		expect:
 		xcode.getSimctl() == '/Applications/Xcode.app/Contents/Developer/usr/bin/simctl'
 	}
+
+	def "get default toolchain directory"() {
+		given:
+		useDefaultXcode()
+		commandRunner.runWithResult("xcodebuild", "-showBuildSettings") >>  "foo=bar"
+
+		expect:
+		xcode.getToolchainDirectory() == "/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain"
+	}
+
+
+	def "get default toolchain directory from build settings"() {
+		given:
+		useDefaultXcode()
+		File buildSettings = new File("src/test/Resource/xcodebuild-showBuildSettings.txt");
+		commandRunner.runWithResult("xcodebuild", "-showBuildSettings") >>  FileUtils.readFileToString(buildSettings)
+
+		expect:
+		xcode.getToolchainDirectory() == "/Applications/Xcode.app/Contents/Developer/Toolchains/Swift_2.3.xctoolchain"
+	}
 }